+++ CHANGELOG.md
@@ -2,7 +2,11 @@ version: 0.3.2
 ---------------
 
 ### Changed
+* Dropped Python 3.4 support
 * `__submodules__` can now be a dictionary of submodule names that point to submodule attributes.
+
+### Fixed
+* Fixed issues in lazy init
 
 
 version: 0.3.1 - released 2020-02-14